[SyntaxSupport]: Fixing some problems.

+++ lib/MiddleEnd/IR/Value.cpp
@@ -23,7 +23,7 @@ std::string Constant::ValueString() const
     if (IsFPConst())
         return std::to_string(std::get<double>(Val));
     else
-        return std::to_string(std::get<uint64_t>(Val));
+        return std::to_string((int64_t)std::get<uint64_t>(Val));
 }
